// packages/test-force_optimize_speed-long_type_string/gen/msg-annotated.ts
export interface FieldUiBehaviour {
    /**
     * @generated from protobuf field: string label = 2;
     */
    label: string;
    /**
     * @generated from protobuf field: bool required = 3;
     */
    required: boolean;
    /**
     * @generated from protobuf field: spec.FieldUiBehaviour.DataSource autocomplete = 4;
     */
    autocomplete?: FieldUiBehaviour_DataSource;
}
/**
 * @generated from protobuf message spec.FieldUiBehaviour.DataSource
 */
export interface FieldUiBehaviour_DataSource {
    /**
     * @generated from protobuf field: string service_name = 1;
     */
    serviceName: string;
    /**
     * @generated from protobuf field: string method_name = 2;
     */
    methodName: string;
    /**
     * @generated from protobuf field: string request_field_name = 3;
     */
    requestFieldName: string;
}
/**
 * shows the usage of custom field options
 *
 * @generated from protobuf message spec.AnnotatedMessage
 */
export interface AnnotatedMessage {
    /**
     * @generated from protobuf field: int32 not_annotated = 1;
     */
    notAnnotated: number;
    /**
     * annotated with the "spec.field_ui" field option
     *
     * @generated from protobuf field: string user_name = 2;
     */
    userName: string;
    /**
     * scalar annotations
     *
     * @generated from protobuf field: int32 ann_scalar = 3;
     */
    annScalar: number;
    /**
     * @generated from protobuf field: int32 ann_repeated_scalar = 4;
     */
    annRepeatedScalar: number;
    /**
     * enum annotations
     *
     * @generated from protobuf field: int32 ann_enum = 5;
     */
    annEnum: number;
    /**
     * @generated from protobuf field: int32 ann_enum_zero = 6;
     */
    annEnumZero: number;
    /**
     * @generated from protobuf field: int32 ann_repeated_enum = 7;
     */
    annRepeatedEnum: number;
    /**
     * annotated with a field option defined right in the message
     *
     * @generated from protobuf field: int32 ann_local = 10;
     */
    annLocal: number;
}
// ...
class FieldUiBehaviour$Type extends MessageType<FieldUiBehaviour> {
    constructor() {
        super("spec.FieldUiBehaviour", [
            { no: 2, name: "label", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "required", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 4, name: "autocomplete", kind: "message", T: () => FieldUiBehaviour_DataSource }
        ]);
    }
    create(value?: PartialMessage<FieldUiBehaviour>): FieldUiBehaviour {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.label = "";
        message.required = false;
        if (value !== undefined)
            reflectionMergePartial<FieldUiBehaviour>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: FieldUiBehaviour): FieldUiBehaviour {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string label */ 2:
                    message.label = reader.string();
                    break;
                case /* bool required */ 3:
                    message.required = reader.bool();
                    break;
                case /* spec.FieldUiBehaviour.DataSource autocomplete */ 4:
                    message.autocomplete = FieldUiBehaviour_DataSource.internalBinaryRead(reader, reader.uint32(), options, message.autocomplete);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: FieldUiBehaviour, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string label = 2; */
        if (message.label !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.label);
        /* bool required = 3; */
        if (message.required !== false)
            writer.tag(3, WireType.Varint).bool(message.required);
        /* spec.FieldUiBehaviour.DataSource autocomplete = 4; */
        if (message.autocomplete)
            FieldUiBehaviour_DataSource.internalBinaryWrite(message.autocomplete, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message spec.FieldUiBehaviour
 */
export const FieldUiBehaviour = new FieldUiBehaviour$Type();
// @generated message type with reflection information, may provide speed optimized methods
class FieldUiBehaviour_DataSource$Type extends MessageType<FieldUiBehaviour_DataSource> {
    constructor() {
        super("spec.FieldUiBehaviour.DataSource", [
            { no: 1, name: "service_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "method_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "request_field_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<FieldUiBehaviour_DataSource>): FieldUiBehaviour_DataSource {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.serviceName = "";
        message.methodName = "";
        message.requestFieldName = "";
        if (value !== undefined)
            reflectionMergePartial<FieldUiBehaviour_DataSource>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: FieldUiBehaviour_DataSource): FieldUiBehaviour_DataSource {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string service_name */ 1:
                    message.serviceName = reader.string();
                    break;
                case /* string method_name */ 2:
                    message.methodName = reader.string();
                    break;
                case /* string request_field_name */ 3:
                    message.requestFieldName = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: FieldUiBehaviour_DataSource, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string service_name = 1; */
        if (message.serviceName !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.serviceName);
        /* string method_name = 2; */
        if (message.methodName !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.methodName);
        /* string request_field_name = 3; */
        if (message.requestFieldName !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.requestFieldName);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message spec.FieldUiBehaviour.DataSource
 */
export const FieldUiBehaviour_DataSource = new FieldUiBehaviour_DataSource$Type();
// ...
export const AnnotatedMessage = new AnnotatedMessage$Type();